+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/_container_service_client.py
@@ -111,12 +111,16 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :mod:`v2024_05_01.models<azure.mgmt.containerservice.v2024_05_01.models>`
            * 2024-05-02-preview: :mod:`v2024_05_02_preview.models<azure.mgmt.containerservice.v2024_05_02_preview.models>`
+           * 2024-06-02-preview: :mod:`v2024_06_02_preview.models<azure.mgmt.containerservice.v2024_06_02_preview.models>`
         """
         if api_version == '2024-05-01':
             from .v2024_05_01 import models
             return models
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview import models
+            return models
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview import models
             return models
         raise ValueError("API version {} is not available".format(api_version))
 
@@ -126,12 +130,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2024_05_01.operations.AgentPoolsOperations>`
            * 2024-05-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.AgentPoolsOperations>`
+           * 2024-06-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.AgentPoolsOperations>`
         """
         api_version = self._get_api_version('agent_pools')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import AgentPoolsOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import AgentPoolsOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import AgentPoolsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'agent_pools'".format(api_version))
         self._config.api_version = api_version
@@ -142,10 +149,13 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         """Instance depends on the API version:
 
            * 2024-05-02-preview: :class:`LoadBalancersOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.LoadBalancersOperations>`
+           * 2024-06-02-preview: :class:`LoadBalancersOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.LoadBalancersOperations>`
         """
         api_version = self._get_api_version('load_balancers')
         if api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import LoadBalancersOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import LoadBalancersOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'load_balancers'".format(api_version))
         self._config.api_version = api_version
@@ -156,10 +166,13 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         """Instance depends on the API version:
 
            * 2024-05-02-preview: :class:`MachinesOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.MachinesOperations>`
+           * 2024-06-02-preview: :class:`MachinesOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.MachinesOperations>`
         """
         api_version = self._get_api_version('machines')
         if api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import MachinesOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import MachinesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'machines'".format(api_version))
         self._config.api_version = api_version
@@ -171,12 +184,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2024_05_01.operations.MaintenanceConfigurationsOperations>`
            * 2024-05-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.MaintenanceConfigurationsOperations>`
+           * 2024-06-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.MaintenanceConfigurationsOperations>`
         """
         api_version = self._get_api_version('maintenance_configurations')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import MaintenanceConfigurationsOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import MaintenanceConfigurationsOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import MaintenanceConfigurationsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'maintenance_configurations'".format(api_version))
         self._config.api_version = api_version
@@ -187,10 +203,13 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         """Instance depends on the API version:
 
            * 2024-05-02-preview: :class:`ManagedClusterSnapshotsOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.ManagedClusterSnapshotsOperations>`
+           * 2024-06-02-preview: :class:`ManagedClusterSnapshotsOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.ManagedClusterSnapshotsOperations>`
         """
         api_version = self._get_api_version('managed_cluster_snapshots')
         if api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import ManagedClusterSnapshotsOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import ManagedClusterSnapshotsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'managed_cluster_snapshots'".format(api_version))
         self._config.api_version = api_version
@@ -202,12 +221,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2024_05_01.operations.ManagedClustersOperations>`
            * 2024-05-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.ManagedClustersOperations>`
+           * 2024-06-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.ManagedClustersOperations>`
         """
         api_version = self._get_api_version('managed_clusters')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import ManagedClustersOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import ManagedClustersOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import ManagedClustersOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'managed_clusters'".format(api_version))
         self._config.api_version = api_version
@@ -218,10 +240,13 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         """Instance depends on the API version:
 
            * 2024-05-02-preview: :class:`OperationStatusResultOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.OperationStatusResultOperations>`
+           * 2024-06-02-preview: :class:`OperationStatusResultOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.OperationStatusResultOperations>`
         """
         api_version = self._get_api_version('operation_status_result')
         if api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import OperationStatusResultOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import OperationStatusResultOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'operation_status_result'".format(api_version))
         self._config.api_version = api_version
@@ -233,12 +258,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`Operations<azure.mgmt.containerservice.v2024_05_01.operations.Operations>`
            * 2024-05-02-preview: :class:`Operations<azure.mgmt.containerservice.v2024_05_02_preview.operations.Operations>`
+           * 2024-06-02-preview: :class:`Operations<azure.mgmt.containerservice.v2024_06_02_preview.operations.Operations>`
         """
         api_version = self._get_api_version('operations')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import Operations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import Operations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import Operations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'operations'".format(api_version))
         self._config.api_version = api_version
@@ -250,12 +278,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2024_05_01.operations.PrivateEndpointConnectionsOperations>`
            * 2024-05-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.PrivateEndpointConnectionsOperations>`
+           * 2024-06-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.PrivateEndpointConnectionsOperations>`
         """
         api_version = self._get_api_version('private_endpoint_connections')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import PrivateEndpointConnectionsOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import PrivateEndpointConnectionsOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import PrivateEndpointConnectionsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'private_endpoint_connections'".format(api_version))
         self._config.api_version = api_version
@@ -267,12 +298,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2024_05_01.operations.PrivateLinkResourcesOperations>`
            * 2024-05-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.PrivateLinkResourcesOperations>`
+           * 2024-06-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.PrivateLinkResourcesOperations>`
         """
         api_version = self._get_api_version('private_link_resources')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import PrivateLinkResourcesOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import PrivateLinkResourcesOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import PrivateLinkResourcesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'private_link_resources'".format(api_version))
         self._config.api_version = api_version
@@ -284,12 +318,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2024_05_01.operations.ResolvePrivateLinkServiceIdOperations>`
            * 2024-05-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
+           * 2024-06-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
         """
         api_version = self._get_api_version('resolve_private_link_service_id')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'resolve_private_link_service_id'".format(api_version))
         self._config.api_version = api_version
@@ -301,12 +338,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2024_05_01.operations.SnapshotsOperations>`
            * 2024-05-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.SnapshotsOperations>`
+           * 2024-06-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.SnapshotsOperations>`
         """
         api_version = self._get_api_version('snapshots')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import SnapshotsOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import SnapshotsOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import SnapshotsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'snapshots'".format(api_version))
         self._config.api_version = api_version
@@ -318,12 +358,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2024_05_01.operations.TrustedAccessRoleBindingsOperations>`
            * 2024-05-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.TrustedAccessRoleBindingsOperations>`
+           * 2024-06-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.TrustedAccessRoleBindingsOperations>`
         """
         api_version = self._get_api_version('trusted_access_role_bindings')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import TrustedAccessRoleBindingsOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import TrustedAccessRoleBindingsOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import TrustedAccessRoleBindingsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'trusted_access_role_bindings'".format(api_version))
         self._config.api_version = api_version
@@ -335,12 +378,15 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2024-05-01: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2024_05_01.operations.TrustedAccessRolesOperations>`
            * 2024-05-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2024_05_02_preview.operations.TrustedAccessRolesOperations>`
+           * 2024-06-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2024_06_02_preview.operations.TrustedAccessRolesOperations>`
         """
         api_version = self._get_api_version('trusted_access_roles')
         if api_version == '2024-05-01':
             from .v2024_05_01.operations import TrustedAccessRolesOperations as OperationClass
         elif api_version == '2024-05-02-preview':
             from .v2024_05_02_preview.operations import TrustedAccessRolesOperations as OperationClass
+        elif api_version == '2024-06-02-preview':
+            from .v2024_06_02_preview.operations import TrustedAccessRolesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'trusted_access_roles'".format(api_version))
         self._config.api_version = api_version